+++ packages/ui/src/app/components/AppPage.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { Page } from '../../common/components/page/Page'
 import { PageContent } from '../../common/components/page/PageContent'
 import { PageHeader } from '../../common/components/page/PageHeader'
-import { PageTabs } from '../../common/components/page/PageTabs'
+import { PageTabs, TabProps } from '../../common/components/page/PageTabs'
 import { PageTitle } from '../../common/components/page/PageTitle'
 import { Breadcrumbs } from '../../common/components/page/Sidebar/Breadcrumbs/Breadcrumbs'
 
@@ -13,7 +13,7 @@ import { SideBar } from './SideBar'
 interface AppPageProps {
   crumbs: { href: string; text: string }[]
   pageTitle: string
-  tabs: { onClick: () => void; active: boolean; title: string }[]
+  tabs: TabProps[]
   children: ReactNode
 }
