+++ packages/ui/src/app/pages/Profile/MyAccounts/MyAccounts.tsx
@@ -1,20 +1,32 @@
 import React from 'react'
+import { useHistory, useRouteMatch } from 'react-router-dom'
 
 import { useTotalBalances } from '../../../../accounts/hooks/useTotalBalances'
 import { Page } from '../../../../common/components/page/Page'
 import { PageContent } from '../../../../common/components/page/PageContent'
 import { PageHeader } from '../../../../common/components/page/PageHeader'
+import { PageTabs } from '../../../../common/components/page/PageTabs'
 import { PageTitle } from '../../../../common/components/page/PageTitle'
 import { Breadcrumbs } from '../../../../common/components/page/Sidebar/Breadcrumbs/Breadcrumbs'
 import { Statistics } from '../../../../common/components/statistics/Stats'
 import { SideBar } from '../../../components/SideBar'
 import { MyProfile, MyProfileContent } from '../Components'
-import { MyProfileTabs } from '../MyProfileTabs'
 
 import { Accounts } from './Accounts'
 
 export function MyAccounts() {
   const { total, transferable, locked, recoverable } = useTotalBalances()
+  const history = useHistory()
+  const isProfile = !!useRouteMatch({
+    exact: true,
+    path: '/profile',
+  })
+  const isMembers = !!useRouteMatch('/profile/memberships')
+
+  const tabs = [
+    { title: 'My accounts', active: isProfile, onClick: () => history.push('/profile') },
+    { title: 'My memberships', active: isMembers, onClick: () => history.push('/profile/memberships') },
+  ]
 
   return (
     <Page>
@@ -29,7 +41,7 @@ export function MyAccounts() {
         <MyProfile>
           <PageHeader>
             <PageTitle>My profile</PageTitle>
-            <MyProfileTabs />
+            <PageTabs tabs={tabs} />
             <Statistics
               stats={[
                 { title: 'Total balance', helperText: 'Lorem fishy', value: total },
