+++ packages/ui/src/app/pages/Profile/MyMemberships/MyMemberships.tsx
@@ -1,35 +1,31 @@
 import React from 'react'
+import { useHistory, useRouteMatch } from 'react-router-dom'
 
-import { Page } from '../../../../common/components/page/Page'
-import { PageContent } from '../../../../common/components/page/PageContent'
-import { PageHeader } from '../../../../common/components/page/PageHeader'
-import { PageTitle } from '../../../../common/components/page/PageTitle'
-import { Breadcrumbs } from '../../../../common/components/page/Sidebar/Breadcrumbs/Breadcrumbs'
-import { SideBar } from '../../../components/SideBar'
-import { MyProfile, MyMembershipsContent } from '../Components'
-import { MyProfileTabs } from '../MyProfileTabs'
+import { AppPage } from '../../../components/AppPage'
 
 import { Memberships } from './Memberships'
 
-export const MyMemberships = () => (
-  <Page>
-    <SideBar />
-    <PageContent>
-      <Breadcrumbs
-        crumbs={[
-          { href: '#', text: 'My Profile' },
-          { href: '#', text: 'My Memberships' },
-        ]}
-      />
-      <MyProfile>
-        <PageHeader>
-          <PageTitle>My profile</PageTitle>
-          <MyProfileTabs />
-        </PageHeader>
-        <MyMembershipsContent>
-          <Memberships />
-        </MyMembershipsContent>
-      </MyProfile>
-    </PageContent>
-  </Page>
-)
+export const MyMemberships = () => {
+  const history = useHistory()
+  const isProfile = !!useRouteMatch({
+    exact: true,
+    path: '/profile',
+  })
+  const isMembers = !!useRouteMatch('/profile/memberships')
+
+  const tabs = [
+    { title: 'My accounts', active: isProfile, onClick: () => history.push('/profile') },
+    { title: 'My memberships', active: isMembers, onClick: () => history.push('/profile/memberships') },
+  ]
+
+  const crumbs = [
+    { href: '#', text: 'My Profile' },
+    { href: '#', text: 'My Memberships' },
+  ]
+
+  return (
+    <AppPage pageTitle="My Profile" crumbs={crumbs} tabs={tabs}>
+      <Memberships />
+    </AppPage>
+  )
+}
